requirements.txt
python-telegram-bot==20.6
pymongo==4.6.0
dnspython==2.4.2
certifi==2023.7.22
Flask==3.0.0
psutil==5.9.5
GitPython==3.1.40
httpx==0.25.0
gTTS==2.4.0
edge-tts==6.1.9
langdetect==1.0.9